Made it all functional

+++ include/board.h
@@ -6,7 +6,7 @@
 
 /* Type for each cell of board */
 struct board_cell_t {
-    size_t x, y;
+    long x, y;
     char val;
     char uncovered;
     char flag;
@@ -14,17 +14,17 @@ struct board_cell_t {
 
 /* Type for board itself, so more than one can be made and it can be saved */
 struct board_t {
-    size_t w, h;
+    long w, h;
     int nmines;
     struct board_cell_t *b;
 };
 
 /* New board with given size */
-struct board_t *board_new(size_t w, size_t h, size_t sx, size_t sy, int nmines);
+struct board_t *board_new(long w, long h, long sx, long sy, int nmines);
 /* Uncover specific location on board and uncover neighbors of it
  * Returns 0 on uncovering cell and neighbors, 1 on cell not being uncoverable,
  * 2 on cell being mine */
-int board_uncover(struct board_t *board, size_t x, size_t y);
+int board_uncover(struct board_t *board, long x, long y);
 /* Destructor */
 void board_del(struct board_t *board);
